requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
gevent==1.0.2
greenlet==0.4.7
itsdangerous==0.24
locustio==0.7.3
msgpack-python==0.4.6
pyzmq==14.6.0
requests==2.7.0
statsd==3.1
wsgiref==0.1.2
PyJWT==1.5.0
arrow==0.10.0
plumbum==1.6.3